重构代码，将代理管理相关的类型和逻辑从 proxyManager 移动到 controller，简化组件中的导入路径，提升代码可读性和维护性。
